emnist/requirements.txt
matplotlib
pygame-ce
torch
torchmetrics
torchvision
tqdm
